Test if git opens the commit message template
When committing without -m message, as in "git commit", will lead to
this page. It seems Git has a natural width of 74 characters or so.
That's pleasant enough a width at least. Well this is interesting and
just a test so I'm having fun with it.
What the hell is a ticket: X by the way?

[ticket: X]